refactored app directory structure
