+++ docker-compose-dev.yml
@@ -10,6 +10,6 @@ services:
     ports: 
       - "5000:5000"
     env_file:
-      - ./.env
+      - ./todo_app/.env
     volumes:
-      - ./:/code/:cached
+      - ./todo_app:/code/todo_app:cached
