+++ entrypoint-dev.sh
@@ -1,3 +1,4 @@
 #!/bin/bash
 
+cd todo_app
 poetry run flask run --host=0.0.0.0
