+++ entrypoint-prod.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-poetry run gunicorn --config gunicorn_config.py 'app:create_app()'
+poetry run gunicorn --config gunicorn_config.py 'todo_app.app:create_app()'
