+++ todo_app/tests/test_todos.py
@@ -1,7 +1,7 @@
 import pytest
-from todo_item import todo_item
-from todo_status import todo_status
-from view_model import view_model
+from todo_app.todo_item import todo_item
+from todo_app.todo_status import todo_status
+from todo_app.view_model import view_model
 from datetime import *
